refactor: refactor package to MVVM pattern
