+++ app/src/main/AndroidManifest.xml
@@ -17,19 +17,19 @@
         android:theme="@style/Theme.Yomakase"
         tools:targetApi="31">
         <activity
-            android:name=".JoinOwnerMemberActivity"
+            android:name=".view.JoinOwnerMemberActivity"
             android:exported="false" />
         <activity
-            android:name=".JoinGeneralMemberActivity"
+            android:name=".view.JoinGeneralMemberActivity"
             android:exported="false" />
         <activity
-            android:name=".JoinOptionActivity"
+            android:name=".view.JoinOptionActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".view.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".view.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
