+++ app/src/main/java/com/example/yomakase/model/ClosedDay.kt
@@ -1,4 +1,4 @@
-package com.example.yomakase
+package com.example.yomakase.model
 
 data class ClosedDay(
     val day: String
