+++ app/src/main/java/com/example/yomakase/model/Facility.kt
@@ -1,4 +1,4 @@
-package com.example.yomakase
+package com.example.yomakase.model
 
 data class Facility(
     val name: String
