+++ app/src/main/java/com/example/yomakase/model/Price.kt
@@ -1,4 +1,4 @@
-package com.example.yomakase
+package com.example.yomakase.model
 
 data class Price(
     val day: String,
