+++ app/src/main/java/com/example/yomakase/view/LoginActivity.kt
@@ -1,7 +1,8 @@
-package com.example.yomakase
+package com.example.yomakase.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.yomakase.R
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
