+++ app/src/main/java/com/example/yomakase/view/SplashActivity.kt
@@ -1,10 +1,11 @@
-package com.example.yomakase
+package com.example.yomakase.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import com.example.yomakase.R
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
