+++ app/src/main/res/layout/activity_find_account.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".FindAccountActivity">
+        tools:context=".view.FindAccountActivity">
 
         <ImageButton
             android:id="@+id/btnBack"
