+++ app/src/main/res/layout/activity_join_general_member.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".JoinGeneralMemberActivity">
+        tools:context=".view.JoinGeneralMemberActivity">
 
         <ImageButton
             android:id="@+id/btnBack"
