+++ app/src/main/res/layout/activity_join_option.xml
@@ -11,7 +11,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".JoinOptionActivity"
+        tools:context=".view.JoinOptionActivity"
         android:backgroundTint="@color/theme_pink">
 
         <ImageButton
