+++ app/src/main/res/layout/activity_join_owner_member.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".JoinOwnerMemberActivity"
+        tools:context=".view.JoinOwnerMemberActivity"
         android:backgroundTint="@color/white">
 
         <ImageButton
