+++ app/src/main/res/layout/activity_login.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".LoginActivity">
+        tools:context=".view.LoginActivity">
 
         <!-- 로고 이미지뷰 -->
