+++ app/src/main/res/layout/rv_item_dlg_facility.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="facility"
-            type="com.example.yomakase.DialogFacility" />
+            type="com.example.yomakase.model.DialogFacility" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
